amended routes upon teacher's suggestions
