+++ app/controllers/freelancer/educations_controller.rb
