+++ app/controllers/users_controller.rb
@@ -1,2 +1,3 @@
 class UsersController < ApplicationController
+
 end
