+++ app/models/review.rb
@@ -5,3 +5,4 @@ class Review < ApplicationRecord
 
   validates :score, numericality: { greater_than_or_equal_to: 0, less_than_or_equal_to: 5 }
 end
+
