+++ app/views/freelancer/profiles/index.html.erb
@@ -37,7 +37,7 @@
 
             <!-- Bouton voir le profil -->
             <div class="mt-3">
-              <%= link_to "Voir le profil", profile_path(profile), class: "btn btn-outline-primary" %>
+              <%= link_to "Voir le profil", freelancer_profile_path(profile), class: "btn btn-outline-primary" %>
             </div>
           </div>
         </div>
