+++ app/views/freelancer/shared_projects/show.html.erb
@@ -18,4 +18,4 @@
   <p class="alert alert-danger">You do not have any project yet.</p>
 <% end %>
 <!-- You can also display a link to go back to the previous page -->
-<%= link_to 'Back to My Projects', user_shared_projects_path, class: 'btn btn-secondary mt-3' %>
+<%= link_to 'Back to My Projects', freelancer_shared_projects_path, class: 'btn btn-secondary mt-3' %>
